move brand to global
